30/10/25 (morning)

+++ app/schemas/material_schema.py
@@ -1,10 +1,11 @@
 from pydantic import BaseModel
+from decimal import Decimal
 
 class CreateMaterial(BaseModel):
     name: str
     short_code: str
     hsn_code: str
-    basic_duty_pr: float
-    social_duty_pr: float
-    igst_pr: float
+    basic_duty_pr: Decimal
+    social_duty_pr: Decimal
+    igst_pr: Decimal
